+++ CHANGELOG.md
@@ -33,3 +33,4 @@
 - PR #95 Faster import time, segregated in-notebook asset loading to save import costs, updated tests
 - PR #100 Bug Fixes - Added NaN value handling for custom bin computations in numba kernels
 - PR #104 Bug Fixes - fixed naming issue for geo column for choropleth3d charts, which did not allow all-small-caps names
+- PR #112 - updated bokeh dependecy to be 1.* instead of >1